update: move utils from graph to generic
