+++ test/UnitTest.cpp
@@ -15,6 +15,7 @@
 #include "TestImage.hpp"
 #include "TestBoolean.hpp"
 #include "TestGraph.hpp"
+#include "TestUtils.hpp"
 using namespace boost::unit_test;
 
 extern test_suite * create_math_test_suite();
@@ -26,6 +27,7 @@ extern test_suite * create_circuit_test_suite();
 extern test_suite * create_image_test_suite();
 extern test_suite * create_boolean_test_suite();
 extern test_suite * create_graph_test_suite();
+extern test_suite * create_utils_test_suite();
 
 void t_additional()
 {
@@ -45,6 +47,7 @@ init_unit_test_suite(int argc, char* argv[])
     framework::master_test_suite().add(create_image_test_suite());
     framework::master_test_suite().add(create_boolean_test_suite());
     framework::master_test_suite().add(create_graph_test_suite());
+    framework::master_test_suite().add(create_utils_test_suite());
     framework::master_test_suite().add(BOOST_TEST_CASE(&t_additional));
     return 0;
 }
